Move proptest generating code to proptest module
