+++ mqtt/mqtt-broker/Cargo.toml
@@ -19,6 +19,7 @@ humantime = "2.0"
 humantime-serde = "1.0"
 lazy_static = "1.4"
 native-tls = "0.2"
+proptest = { version = "0.9", optional = true }
 rand = { version = "0.7", optional = true }
 regex = "1"
 serde = { version = "1.0", features = ["derive", "rc"] }
