+++ mqtt/mqtt-broker/src/broker.rs
@@ -1042,34 +1042,18 @@ pub(crate) mod tests {
     use bytes::Bytes;
     use futures_util::future::FutureExt;
     use matches::assert_matches;
-    use proptest::collection::{hash_map, vec};
-    use proptest::prelude::*;
-    use tokio::sync::mpsc::error::TryRecvError;
-    use tokio::sync::mpsc::{self, UnboundedReceiver};
+    use tokio::sync::mpsc::{self, error::TryRecvError, UnboundedReceiver};
     use uuid::Uuid;
 
     use mqtt3::{proto, PROTOCOL_LEVEL, PROTOCOL_NAME};
 
     use crate::{
         auth::{Activity, AuthenticateError, AuthorizeError, Operation},
-        broker::{BrokerBuilder, BrokerHandle, BrokerState, SessionError},
+        broker::{BrokerBuilder, BrokerHandle, SessionError},
         error::Error,
-        session::{tests::arb_session_state, Session},
-        tests::{arb_publication, arb_topic},
+        session::Session,
         AuthId, ClientEvent, ClientId, ConnReq, ConnectionHandle, Message,
     };
-
-    prop_compose! {
-        pub fn arb_broker_state()(
-            retained in hash_map(arb_topic(), arb_publication(), 0..20),
-            sessions in vec(arb_session_state(), 0..10),
-        ) -> BrokerState {
-            BrokerState {
-                retained,
-                sessions,
-            }
-        }
-    }
 
     fn connection_handle() -> ConnectionHandle {
         let id = Uuid::new_v4();
