+++ mqtt/mqtt-broker/src/lib.rs
@@ -42,6 +42,9 @@ pub use crate::snapshot::{Snapshotter, StateSnapshotHandle};
 pub use crate::subscription::{Segment, Subscription, TopicFilter};
 pub use crate::transport::TransportBuilder;
 
+#[cfg(any(test, feature = "proptest"))]
+pub mod proptest;
+
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct ClientId(Arc<String>);
 
@@ -188,91 +191,4 @@ pub enum SystemEvent {
 pub enum Message {
     Client(ClientId, ClientEvent),
     System(SystemEvent),
-}
-
-#[cfg(test)]
-pub(crate) mod tests {
-    use std::sync::Arc;
-
-    use bytes::Bytes;
-    use proptest::collection::vec;
-    use proptest::num;
-    use proptest::prelude::*;
-
-    use mqtt3::proto;
-
-    use crate::{subscription::tests::arb_qos, ClientId, Publish};
-
-    pub fn arb_clientid() -> impl Strategy<Value = ClientId> {
-        "[a-zA-Z0-9]{1,23}".prop_map(|s| ClientId(Arc::new(s)))
-    }
-
-    pub fn arb_packet_identifier() -> impl Strategy<Value = proto::PacketIdentifier> {
-        (1_u16..=u16::max_value())
-            .prop_map(|i| proto::PacketIdentifier::new(i).expect("packet identifier failed"))
-    }
-
-    pub fn arb_topic() -> impl Strategy<Value = String> {
-        "\\PC+(/\\PC+)*"
-    }
-
-    pub fn arb_payload() -> impl Strategy<Value = Bytes> {
-        vec(num::u8::ANY, 0..1024).prop_map(Bytes::from)
-    }
-
-    prop_compose! {
-        pub fn arb_publication()(
-            topic_name in arb_topic(),
-            qos in arb_qos(),
-            retain in proptest::bool::ANY,
-            payload in arb_payload(),
-        ) -> proto::Publication {
-            proto::Publication {
-                topic_name,
-                qos,
-                retain,
-                payload,
-            }
-        }
-    }
-
-    pub fn arb_pidq() -> impl Strategy<Value = proto::PacketIdentifierDupQoS> {
-        prop_oneof![
-            Just(proto::PacketIdentifierDupQoS::AtMostOnce),
-            (arb_packet_identifier(), proptest::bool::ANY)
-                .prop_map(|(id, dup)| proto::PacketIdentifierDupQoS::AtLeastOnce(id, dup)),
-            (arb_packet_identifier(), proptest::bool::ANY)
-                .prop_map(|(id, dup)| proto::PacketIdentifierDupQoS::ExactlyOnce(id, dup)),
-        ]
-    }
-
-    prop_compose! {
-        pub fn arb_proto_publish()(
-            pidq in arb_pidq(),
-            retain in proptest::bool::ANY,
-            topic_name in arb_topic(),
-            payload in arb_payload(),
-        ) -> proto::Publish {
-            proto::Publish {
-                packet_identifier_dup_qos: pidq,
-                retain,
-                topic_name,
-                payload,
-            }
-        }
-    }
-
-    pub fn arb_publish() -> impl Strategy<Value = Publish> {
-        prop_oneof![
-            (arb_packet_identifier(), arb_proto_publish())
-                .prop_map(|(id, publish)| Publish::QoS0(id, publish)),
-            (arb_packet_identifier(), arb_proto_publish())
-                .prop_map(|(id, publish)| Publish::QoS12(id, publish)),
-        ]
-    }
-
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
 }
