+++ mqtt/mqtt-broker/src/subscription.rs
@@ -163,61 +163,9 @@ impl FromStr for TopicFilter {
 pub(crate) mod tests {
     use std::str::FromStr;
 
-    use proptest::bool;
-    use proptest::collection::vec;
     use proptest::prelude::*;
 
-    use mqtt3::proto;
-
-    use crate::subscription::{Segment, Subscription, TopicFilter};
-
-    fn arb_segment() -> impl Strategy<Value = Segment> {
-        prop_oneof![
-            "[^+#\0/]+".prop_map(Segment::Level),
-            Just(Segment::SingleLevelWildcard),
-            Just(Segment::MultiLevelWildcard),
-        ]
-    }
-
-    prop_compose! {
-        pub fn arb_topic_filter()(
-            segments in vec(arb_segment(), 1..20),
-            multi in bool::ANY,
-        ) -> TopicFilter {
-            let mut filtered = vec![];
-            for segment in segments {
-                if segment != Segment::MultiLevelWildcard {
-                    filtered.push(segment);
-                }
-            }
-
-            if multi || filtered.is_empty() {
-                filtered.push(Segment::MultiLevelWildcard);
-            }
-
-            TopicFilter::new(filtered)
-        }
-    }
-
-    pub fn arb_qos() -> impl Strategy<Value = proto::QoS> {
-        prop_oneof![
-            Just(proto::QoS::AtMostOnce),
-            Just(proto::QoS::AtLeastOnce),
-            Just(proto::QoS::ExactlyOnce),
-        ]
-    }
-
-    prop_compose! {
-        pub fn arb_subscription()(
-            filter in arb_topic_filter(),
-            max_qos in arb_qos(),
-        ) -> Subscription {
-            Subscription {
-                filter,
-                max_qos,
-            }
-        }
-    }
+    use crate::subscription::{Segment, TopicFilter};
 
     fn filter(segments: Vec<Segment>) -> TopicFilter {
         TopicFilter::new(segments)
@@ -293,7 +241,7 @@ pub(crate) mod tests {
 
     proptest! {
         #[test]
-        fn display_roundtrip(filter in arb_topic_filter()) {
+        fn display_roundtrip(filter in crate::proptest::arb_topic_filter()) {
             let string = filter.to_string();
             prop_assert_eq!(filter, string.parse::<TopicFilter>().unwrap());
         }
